Fully working on Lambda.
Most of the work was in pulling apart the payload and creating a
namedtuple to hold the few needed attributes. Also needed to use
requests to hit a couple of GitHub API endpoints.

This also re-orders looking for matching Tasks to first try to create
one, and then falls back to finding an existing. This should be much
more race condition proof.